renamed files; useLogin implemented
