+++ src/data/INote.ts
@@ -1,15 +1,15 @@
 export interface INote {
-    data:    Data[];
+    data:    INoteData[];
     success: boolean;
     message: null;
 }
 
-export interface Data {
-    userID:    number;
-    noteID:    number;
-    title:     string;
-    content:   string;
-    priority:  number;
-    category:  number;
+export interface INoteData {
+    userId: number;
+    noteId: number;
+    title: string;
+    content: string;
+    priority: number;
+    category: number;
     createdOn: Date;
 }
