--- src/data/INoteData.ts
@@ -1,9 +1,0 @@
-export interface INoteData {
-    userId: string;
-    noteId: number;
-    title: string;
-    content: string;
-    priority: number;
-    category: number;
-    createdOn: Date;
-}
